# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Game-Baza API
  description: 'REST API платформы для сравнения цен на игры. Поддерживает каталог игр, фильтрацию, аутентификацию через JWT-cookie.'
  contact:
    name: Game-Baza Team
    url: 'https://github.com/bembelem/game-baza'
    email: support@game-baza.example.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 0.1.0
servers:
  - url: 'http://localhost:8080'
    description: Локальная разработка
paths:
...
  /users/{user_id}:
    get:
      tags: [ Users ]
      summary: Публичные данные пользователя
      description: Возвращает публичные данные пользователя.
      operationId: getUserById
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Данные профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/me:
    get:
      tags: [Users]
      summary: Мой профиль
      description: Возвращает полный профиль авторизованного пользователя.
      operationId: getMe
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Данные профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPrivate'
        '401':
          description: Токен не передан или не действителен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
    patch:
      tags: [Users]
      summary: Изменить профиль
      description: 'Частично обновляет профиль авторизованного пользователя. Поля передаются опционально, пустые игнорируются.'
      operationId: updateMe
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatch'
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          description: Токен не передан или не действителен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '409':
          description: Email или username уже занят
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
    delete:
      tags: [Users]
      summary: Удалить профиль
      description: Удаляет профиль авторизованного пользователя и сбрасывает cookie.
      operationId: deleteMe
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Профиль удалён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Токен не передан или не действителен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT-токен в HttpOnly cookie. Устанавливается /auth/login и /auth/register.
  schemas:
...
    UserPublic:
      type: object
      title: UserPublic
      description: Публичный профиль пользователя — безопасные поля, отдаётся другим.
      required: [id, username, created_at]
      properties:
        id:
          type: integer
        username:
          type: string
        created_at:
          type: string
          format: date
      example:
        id: 42
        username: john_doe
        created_at: '2024-05-20'
    UserPrivate:
      type: object
      title: UserPrivate
      description: Полный профиль пользователя — отдаётся только владельцу (/users/me).
      required: [id, username, email, birthdate, created_at]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        birthdate:
          type: string
          format: date
        created_at:
          type: string
          format: date
      example:
        id: 42
        username: john_doe
        email: john@example.com
        birthdate: '2000-01-15'
        created_at: '2024-05-20'
    UserPatch:
      type: object
      title: UserPatch
      description: Частичное обновление профиля. Все поля опциональны — присылай только то что меняешь.
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_ ]+$'
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 50
          pattern: '^(?=.*\d)(?=.*[!@#$%^&*()_\-+=\[\]{};:''\",.<>/?\\|`~]).+$'
        birthdate:
          type: string
          format: date
      example:
        username: new_nickname
        birthdate: '2000-01-15'

    # ----- Games -----
...
    MessageResponse:
      type: object
      title: MessageResponse
      required: [message]
      properties:
        message:
          type: string
      example:
        message: Профиль удалён
    ErrorDetailResponse:
      type: object
      title: ErrorDetailResponse
      description: Унифицированный формат всех ошибок API.
      required: [error, message, details, traceId]
      properties:
        error:
          type: string
          description: Машино-читаемый код ошибки (например 2010_AUTHENTICATION_ERROR)
        message:
          type: string
          description: Человеко-читаемое сообщение
        details:
          type: object
          additionalProperties: true
          description: Детали ошибки по полям
        traceId:
          type: string
          description: ID трассировки запроса (X-Request-ID), для дебага
      example:
        error: 2010_AUTHENTICATION_ERROR
        message: Ошибка авторизации.
        details:
          password: Неверный пароль.
        traceId: 9696004a-6866-4a79-bc53-f4383040cba1

  # ----- Error Responses -----
  responses:
    ValidationError:
      description: Ошибка валидации параметров или тела
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetailResponse'
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetailResponse'
...